+++ blueprints/flexberry-model/ModelBlueprint.ts
@@ -29,6 +29,7 @@ export default class ModelBlueprint {
   parentClassName: string;
   parentExternal: boolean;
   className: string;
+  namespace: string;
   projections: string;
   name: string;
   needsAllModels: string;
@@ -52,6 +53,7 @@ export default class ModelBlueprint {
     }
     this.enums = ModelBlueprint.loadEnums(options.metadataDir);
     this.className = model.className;
+    this.namespace = model.nameSpace;
     this.serializerAttrs = this.getSerializerAttrs(model);
     this.offlineSerializerAttrs = this.getOfflineSerializerAttrs(model);
     this.projections = this.getJSForProjections(model, modelsDir);