+++ blueprints/flexberry-model/files/__root__/mixins/regenerated/models/__name__.js
@@ -6,7 +6,15 @@ import <%= enumName %> from '../../../enums/<%= enumImports[enumName] %>';<%
 } %>
 export let Model = Ember.Mixin.create({
 <%= model %>
-});<%if(parentModelName) {%>
+});
+<% if (namespace) { %>
+export function defineNamespace(modelClass) {
+  modelClass.reopenClass({
+    namespace: '<%= namespace %>',
+  });
+}
+<% }
+if(parentModelName) {%>
 export let defineBaseModel = function (modelClass) {
   modelClass.reopenClass({
     _parentModelName: '<%= parentModelName %>'